requirements.txt
requests
atproto
Pillow